requirements.txt
asttokens==2.0.8
av==9.2.0
backcall==0.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.0
fonttools==4.37.4
idna==3.4
ipykernel==6.16.0
ipython==8.5.0
ipywidgets==8.0.2
jedi==0.18.1
joblib==1.2.0
jupyter_client==7.3.5
jupyter-core==4.11.1
jupyterlab-widgets==3.0.3
kaggle==1.5.12
kiwisolver==1.4.4
matplotlib==3.6.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.3
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pydicom==2.3.0
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.4
pyzmq==24.0.1
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
setuptools==63.4.1
six==1.16.0
stack-data==0.5.1
text-unidecode==1.3
threadpoolctl==3.1.0
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
wheel==0.37.1
widgetsnbextension==4.0.3
